# Conflicts:
#	src/main/java/com/example/ijoa_refactoring/controller/UserController.java
#	src/main/java/com/example/ijoa_refactoring/service/UserService.java

+++ src/main/java/com/example/ijoa_refactoring/data/repository/DolbomiRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface DolbomiRepository extends JpaRepository<Dolbomi,Long> {
+public interface DolbomiRepository extends JpaRepository<Dolbomi,Integer> {
 
     Dolbomi findById(String userId);
